"Fix: Sign in page redirection and remodel"
