+++ app/loginUser.styles.ts
@@ -1,0 +1,47 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 60,
+    paddingHorizontal: 24,
+    backgroundColor: '#f9f9f9',
+  },
+  title: {
+    fontSize: 22,
+    fontWeight: 'bold',
+    color: '#D62626',
+    textAlign: 'center',
+    marginBottom: 32,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 8,
+    padding: 12,
+    fontSize: 16,
+    backgroundColor: '#fff',
+    marginBottom: 12,
+  },
+  button: {
+    backgroundColor: '#D62626',
+    padding: 14,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginTop: 16,
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    fontSize: 16,
+  },
+  footerText: {
+    marginTop: 20,
+    textAlign: 'center',
+    color: '#555',
+  },
+  link: {
+    fontWeight: 'bold',
+    color: '#000',
+  },
+});
