+++ app/loginUser.tsx
@@ -1,0 +1,49 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, TouchableOpacity, StyleSheet } from 'react-native';
+import { useRouter } from 'expo-router';
+import { styles } from './login.styles';
+
+export default function LoginPage() {
+  const [email, setEmail] = useState('');
+  const [senha, setSenha] = useState('');
+  const router = useRouter();
+
+  const handleLogin = () => {
+    console.log('Login com:', email, senha);
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Login</Text>
+
+      <TextInput
+        style={styles.input}
+        placeholder="E-mail"
+        placeholderTextColor="rgba(0, 0, 0, 0.4)"
+        keyboardType="email-address"
+        autoCapitalize="none"
+        value={email}
+        onChangeText={setEmail}
+      />
+
+      <TextInput
+        style={styles.input}
+        placeholder="Senha"
+        placeholderTextColor="rgba(0, 0, 0, 0.4)"
+        secureTextEntry
+        value={senha}
+        onChangeText={setSenha}
+      />
+
+      <TouchableOpacity style={styles.button} onPress={handleLogin}>
+        <Text style={styles.buttonText}>Entrar</Text>
+      </TouchableOpacity>
+
+      <TouchableOpacity onPress={() => router.push('/addUser')}>
+        <Text style={styles.footerText}>
+          Ainda não tem conta? <Text style={styles.link}>Cadastre-se!</Text>
+        </Text>
+      </TouchableOpacity>
+    </View>
+  );
+}
